+++ R/01.R
@@ -1,4 +1,0 @@
-.onLoad <- function(libname, pkgname) {
-  options("DataPackageR_interact" = interactive())
-  options("DataPackageR_packagebuilding" = FALSE)
-}