+++ R/dataversion.R
@@ -8,6 +8,7 @@
 #' @param lib.loc \code{character} path to library location.
 #' @seealso \code{\link[utils]{packageVersion}}
 #' @rdname data_version
+#' @returns Object of class 'package_version' and 'numeric_version' specifying the DataVersion of the package
 #' @note \code{dataVersion()} has been renamed to \code{data_version()}
 #' @importFrom utils capture.output file_test package.skeleton packageDescription
 #' @export
@@ -16,13 +17,13 @@
 #' f <- tempdir()
 #' f <- file.path(f,"foo.Rmd")
 #' con <- file(f)
-#' writeLines("```{r}\n tbl = table(sample(1:10,1000,replace=TRUE)) \n```\n",con=con)
+#' writeLines("```{r}\n vec = 1:10 \n```\n",con=con)
 #' close(con)
 #' pname <- basename(tempfile())
 #' datapackage_skeleton(name = pname,
 #'    path=tempdir(),
 #'    force = TRUE,
-#'    r_object_names = "tbl",
+#'    r_object_names = "vec",
 #'    code_files = f)
 #'
 #'    package_build(file.path(tempdir(),pname), install = FALSE)
@@ -96,13 +97,13 @@ dataVersion <- function(pkg, lib.loc = NULL) {
 #' f <- tempdir()
 #' f <- file.path(f, "foo.Rmd")
 #' con <- file(f)
-#' writeLines("```{r}\n tbl = table(sample(1:10,1000,replace=TRUE)) \n```\n",con = con)
+#' writeLines("```{r}\n vec = 1:10 \n```\n",con = con)
 #' close(con)
 #' pname <- basename(tempfile())
 #' datapackage_skeleton(name = pname,
 #'    path=tempdir(),
 #'    force = TRUE,
-#'    r_object_names = "tbl",
+#'    r_object_names = "vec",
 #'    code_files = f)
 #' package_build(file.path(tempdir(),pname), install = FALSE)
 #'