+++ tests/testthat/test-edge-cases.R
@@ -1,9 +1,8 @@
 context("edge cases")
 test_that("package built in different edge cases", {
-  require(futile.logger)
-  DataPackageR:::.multilog_setup(normalizePath(file.path(tempdir(),"test.log"), winslash = "/"))
+  DataPackageR:::.multilog_setup(file.path(tempdir(),"test.log"))
   DataPackageR:::.multilog_thresold(INFO, TRACE)
-  
+
   file <- system.file("extdata", "tests", "subsetCars.Rmd",
     package = "DataPackageR"
   )
@@ -57,7 +56,9 @@ test_that("package built in different edge cases", {
     force = TRUE,
     recursive = TRUE
   )
-  package.skeleton("foo", path = tempdir())
+  test_env <- new.env()
+  assign('test_obj', pi, envir = test_env)
+  package.skeleton("foo", path = tempdir(), environment = test_env)
   suppressWarnings(expect_error(
     DataPackageR:::DataPackageR(
       file.path(tempdir(), "foo")
@@ -88,7 +89,7 @@ test_that("package built in different edge cases", {
   )
 
 
-  package.skeleton("foo", path = tempdir(), force = TRUE)
+  package.skeleton("foo", path = tempdir(), environment = test_env, force = TRUE)
   expect_error(yml_find(file.path(tempdir(), "foo")))
   dir.create(file.path(tempdir(), "foo", "data-raw"))
   unlink(file.path(tempdir(), "foo", "DESCRIPTION"))
@@ -168,10 +169,11 @@ test_that("package built in different edge cases", {
     force = TRUE,
     recursive = TRUE
   )
-  package.skeleton("foo", path = tempdir())
-  DataPackageR:::.multilog_setup(normalizePath(file.path(tempdir(),"test.log"), winslash = "/"))
+  package.skeleton("foo", path = tempdir(), environment = test_env, force = TRUE)
+  DataPackageR:::.multilog_setup(file.path(tempdir(),"test.log"))
   DataPackageR:::.multilog_thresold(INFO, TRACE)
-  
+
+  # data in digest changes while names do not
   suppressWarnings(expect_false({
     DataPackageR:::.compare_digests(
       list(
@@ -181,6 +183,64 @@ test_that("package built in different edge cases", {
       list(
         DataVersion = "1.1.2",
         a = paste0(LETTERS[1:10], collapse = "")
+      )
+    )
+  }))
+
+  # names in digest changes while data do not
+  suppressWarnings(expect_false({
+    DataPackageR:::.compare_digests(
+      list(
+        DataVersion = "1.1.1",
+        a = paste0(letters[1:10], collapse = "")
+      ),
+      list(
+        DataVersion = "1.1.2",
+        b = paste0(letters[1:10], collapse = "")
+      )
+    )
+  }))
+
+  # names in digest nor data changes
+  suppressWarnings(expect_true({
+    DataPackageR:::.compare_digests(
+      list(
+        DataVersion = "1.1.1",
+        a = paste0(letters[1:10], collapse = "")
+      ),
+      list(
+        DataVersion = "1.1.1",
+        a = paste0(letters[1:10], collapse = "")
+      )
+    )
+  }))
+
+  # names in old digest have one more than new
+  suppressWarnings(expect_false({
+    DataPackageR:::.compare_digests(
+      list(
+        DataVersion = "1.1.1",
+        a = paste0(letters[1:10], collapse = ""),
+        b = paste0(LETTERS[1:10], collapse = "")
+      ),
+      list(
+        DataVersion = "1.1.2",
+        a = paste0(letters[1:10], collapse = "")
+      )
+    )
+  }))
+
+  # names in new digest have one more than old
+  suppressWarnings(expect_false({
+    DataPackageR:::.compare_digests(
+      list(
+        DataVersion = "1.1.1",
+        a = paste0(letters[1:10], collapse = "")
+      ),
+      list(
+        DataVersion = "1.1.2",
+        a = paste0(letters[1:10], collapse = ""),
+        b = paste0(LETTERS[1:10], collapse = "")
       )
     )
   }))
@@ -237,7 +297,7 @@ test_that("package built in different edge cases", {
     force = TRUE,
     recursive = TRUE
   )
-  package.skeleton(path = tempdir(), "foo")
+  package.skeleton("foo", path = tempdir(), environment = test_env, force = TRUE)
   dir.create(file.path(tempdir(), "foo", "data-raw"))
   suppressWarnings(
     expect_error(